fix: ts reference issus
