+++ packages/core/src/models/field/derivate/number.field.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { plainToInstance } from 'class-transformer';
-import { FieldType, DbFieldType } from '../constant';
-import { CellValueType, FieldCore } from '../field';
+import { FieldType, DbFieldType, CellValueType } from '../constant';
+import { FieldCore } from '../field';
 import type { NumberFieldOptions } from './number.field';
 import { NumberFieldCore } from './number.field';
